custom hooks moved to src, keys added, value condition in Header added
